fix GridOffset type

+++ client/src/ui.c
@@ -43,7 +43,7 @@ double PlotGridX = 0, PlotGridY = 0, PlotGridXdefault = 64, PlotGridYdefault = 6
 uint32_t CursorCPos = 0, CursorDPos = 0;
 double GraphPixelsPerPoint = 1.f; // How many visual pixels are between each sample point (x axis)
 static bool flushAfterWrite = 0;
-int GridOffset = 0;
+double GridOffset = 0;
 bool GridLocked = false;
 bool showDemod = true;
